requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
attrs==19.1.0
black==19.3b0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cryptography==2.6.1
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
django-rest-knox==4.0.1
djangorestframework==3.9.1
djangorestframework-simplejwt==3.3
freezegun==0.3.11
gcloud==0.17.0
googleapis-common-protos==1.5.8
gunicorn==19.9.0
httplib2==0.12.1
idna==2.8
influxdb==5.2.2
jws==0.1.3
oauth2client==3.0.0
protobuf==3.7.0
psycopg2==2.7
psycopg2-binary==2.8
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
pycryptodome==3.7.3
PyJWT==1.7.1
Pyrebase==3.0.27
python-dateutil==2.7.5
python-jwt==2.0.1
pytz==2018.9
requests==2.21.0
requests-toolbelt==0.7.0
rsa==4.0
simple-http-server==0.1.7
six==1.12.0
toml==0.10.0
urllib3==1.24.1
whitenoise==4.1.2
